+++ CHANGELOG.md
@@ -4,12 +4,13 @@ CHANGELOG
 ## Not released
 
 *   Schema 2.0.0
+*   Cmdlets renamed: `platyPS` prefix removed from the names.
 *   Removed APIs:
     -   `New-PlatyPSModuleFromMaml` (private now)
-    -   `Get-PlatyPSMarkdown` always produces one md file per cmdlet in `-OutputFolder`.
+    -   `New-Markdown` always produces one md file per cmdlet in `-OutputFolder`.
 *   New APIs:    
-    -   `New-PlatyPSCab` - create updatable help Cab files
-    -   `Get-PlatyPSYamlMetadata` - get a hashtable with yaml metadata (new in schema 2.0.0) from the markdown file
+    -   `New-ExternalHelpCab` - create updatable help Cab files
+    -   `Get-MarkdownMetadata` - get a hashtable with yaml metadata (new in schema 2.0.0) from the markdown file
 *   Generate one file per cmdlet is not the default
 
 ## 0.3.1